% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFilter.R
\name{filterInnerN}
\alias{filterInnerN}
\title{Trim reads containing too many N at inner part of the reads}
\usage{
filterInnerN(s = 3L, e = 12L, thrs = 1L, .name = "InnerNFilter")
}
\description{
Trim reads containing too many N at inner part of the reads
}
