% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFilter.R
\name{filterLowComplexity}
\alias{filterLowComplexity}
\title{Trim reads with low complexity}
\usage{
filterLowComplexity(
  threshold = 0.5,
  referenceEntropy = 3.908135,
  .name = "LowComplexityFilter"
)
}
\description{
Trim reads with low complexity
}
