% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFilter.R
\name{filterLowQuality}
\alias{filterLowQuality}
\title{Trim reads with low mean quality of bases}
\usage{
filterLowQuality(threshold = 20, .name = "LowQualityFilter")
}
\description{
Trim reads with low mean quality of bases
}
