% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqW.R
\name{seqW}
\alias{seqW}
\title{Master function wrapping trimming bases, filtering reads, alignment and feature counting sequentially}
\usage{
seqW(
  fileList1,
  fileList2 = NULL,
  alignType = "rna",
  subReadThreads = 3L,
  shortreadRAM = 1e+08,
  genomeRefFile = "./GCF_000001405.26_GRCh38_genomic.fna.gz",
  genomeAnnotFile = "./GCF_000001405.39_GRCh38.p13_genomic.gtf.gz",
  indexBasename = NULL,
  alignPairedOutput = gsub(basename(fileList1), pattern = "_1.*\\\\.fastq\\\\.gz",
    replacement = "\\\\.bam"),
  ...
)
}
\arguments{
\item{fileList1}{full path of one or multiple fastq.gz files to be processed}

\item{fileList2}{default NULL for single-end, full path of the paired fastq.gz files for paired-end}

\item{alignType}{default "rna" for RNAseq and microRNAseq,see other option in vignettes, it will be passed to the "type" argument of SubRead::align}

\item{subReadThreads}{number of threads to be used by SubRead, default 3L}

\item{shortreadRAM}{RAM limit in bytes for ShortRead, default 1e8}

\item{genomeRefFile, genomeAnnotFile}{full path of refseq sequence fastq file to generate alignment index, and its mate .gtf annotation file for feature counting.These files can be downloaded from NCBI,see vignettes}

\item{indexBasename}{if you never build index before, leave as defaul "NULL", otherwise set it as "my_index" and make sure index files are in working dir to speed up by avoiding regenerating index files}

\item{alignPairedOutput}{only works for paired-end input,default is gsub(basename(fileList1),pattern ="_1.*\\.fastq\\.gz",replacement = "\\.bam"), optimize it please if it happens to cause overwriting of output files}
}
\value{
A list contains Date.frame summarizing trimming and filtration of each input fastq.gz file. Generated .bam file will be save to ./bam; feature counts are saved as .rds file in working dir
}
\description{
Master function wrapping trimming bases, filtering reads, alignment and feature counting sequentially
}
\examples{
trimFilterRes<-seqW(fileList1= fl,subReadThreads = 3L,shortreadRAM=1e8)
}
