% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFilter.R
\name{trimAdapter}
\alias{trimAdapter}
\title{Trim adapter sequences out of reads}
\usage{
trimAdapter(
  shortQobj,
  Lpattern = "",
  Rpattern = "",
  max.Lmismatch = rep(0:2, c(6, 3, 100)),
  max.Rmismatch = rep(0:2, c(6, 3, 100)),
  with.Lindels = F,
  with.Rindels = F
)
}
\description{
Trim adapter sequences out of reads
}
